+++ database/migrations/2024_10_09_162528_add_created_by_and_updated_by_to_tables.php
@@ -11,28 +11,32 @@ return new class extends Migration
      */
     public function up()
     {
-        $tables = [
-            'donor_funds'
-        ];
+        if(file_exists(base_path('config/hrms.php'))) {
+            $tables = [
+                'donor_funds'
+            ];
 
-        foreach ($tables as $table) {
-            Schema::table($table, function (Blueprint $table) {
-                $table->integer('created_by')->nullable();
-                $table->integer('updated_by')->nullable();
-            });
+            foreach ($tables as $table) {
+                Schema::table($table, function (Blueprint $table) {
+                    $table->integer('created_by')->nullable();
+                    $table->integer('updated_by')->nullable();
+                });
+            }
         }
     }
 
     public function down()
     {
-        $tables = [
-            'donor_funds',
-        ];
+        if(file_exists(base_path('config/hrms.php'))) {
+            $tables = [
+                'donor_funds',
+            ];
 
-        foreach ($tables as $table) {
-            Schema::table($table, function (Blueprint $table) {
-                $table->dropColumn(['created_by', 'updated_by']);
-            });
+            foreach ($tables as $table) {
+                Schema::table($table, function (Blueprint $table) {
+                    $table->dropColumn(['created_by', 'updated_by']);
+                });
+            }
         }
     }
 };
